# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "REST API for property listings with geospatial search.",
        "title": "Property Marketplace API",
        "contact": {
            "name": "API Support"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/listings": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "listings"
                ],
                "summary": "List all listings",
                "parameters": [
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "Page number",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 20,
                        "description": "Items per page",
                        "name": "per_page",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal_listing_handler.DataListResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/internal_listing_handler.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "listings"
                ],
                "summary": "Create a listing",
                "parameters": [
                    {
                        "description": "Listing payload",
                        "name": "listing",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_listing_handler.CreateListingRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/internal_listing_handler.DataResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/internal_listing_handler.ErrorResponse"
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "schema": {
                            "$ref": "#/definitions/internal_listing_handler.ValidationErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/internal_listing_handler.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/listings/search": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "listings"
                ],
                "summary": "Search listings",
                "parameters": [
                    {
                        "enum": [
                            "rent",
                            "sale",
                            "shortlet"
                        ],
                        "type": "string",
                        "description": "Listing type",
                        "name": "type",
                        "in": "query"
                    },
                    {
                        "type": "number",
                        "description": "Minimum price",
                        "name": "min_price",
                        "in": "query"
                    },
                    {
                        "type": "number",
                        "description": "Maximum price",
                        "name": "max_price",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Number of bedrooms",
                        "name": "bedrooms",
                        "in": "query"
                    },
                    {
                        "type": "number",
                        "description": "Latitude of search origin",
                        "name": "lat",
                        "in": "query"
                    },
                    {
                        "type": "number",
                        "description": "Longitude of search origin",
                        "name": "lng",
                        "in": "query"
                    },
                    {
                        "type": "number",
                        "description": "Search radius in kilometres",
                        "name": "radius_km",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "Page number",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 20,
                        "description": "Items per page",
                        "name": "per_page",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal_listing_handler.DataListResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/internal_listing_handler.ErrorResponse"
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "schema": {
                            "$ref": "#/definitions/internal_listing_handler.ValidationErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/internal_listing_handler.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/listings/{id}": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "listings"
                ],
                "summary": "Get a listing by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Listing UUID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal_listing_handler.DataResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/internal_listing_handler.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/internal_listing_handler.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/internal_listing_handler.ErrorResponse"
                        }
                    }
                }
            },
            "put": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "listings"
                ],
                "summary": "Update a listing",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Listing UUID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Updated listing payload",
                        "name": "listing",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_listing_handler.UpdateListingRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal_listing_handler.DataResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/internal_listing_handler.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/internal_listing_handler.ErrorResponse"
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "schema": {
                            "$ref": "#/definitions/internal_listing_handler.ValidationErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/internal_listing_handler.ErrorResponse"
                        }
                    }
                }
            },
            "delete": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "listings"
                ],
                "summary": "Delete a listing",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Listing UUID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/internal_listing_handler.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/internal_listing_handler.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/internal_listing_handler.ErrorResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "github_com_davidakpele_property-marketplace_pkg_pagination.Meta": {
            "type": "object",
            "properties": {
                "page": {
                    "type": "integer"
                },
                "per_page": {
                    "type": "integer"
                },
                "total_items": {
                    "type": "integer"
                },
                "total_pages": {
                    "type": "integer"
                }
            }
        },
...
        "internal_listing_handler.CreateListingRequest": {
            "type": "object",
            "required": [
                "address",
                "agent_id",
                "latitude",
                "longitude",
                "price",
                "title",
                "type"
            ],
            "properties": {
                "address": {
                    "type": "string",
                    "maxLength": 500,
                    "minLength": 3,
                    "example": "14 Admiralty Way, Lekki Phase 1"
                },
                "agent_id": {
                    "type": "string",
                    "example": "a1000000-0000-0000-0000-000000000001"
                },
                "bedrooms": {
                    "type": "integer",
                    "minimum": 0,
                    "example": 3
                },
                "description": {
                    "type": "string",
                    "example": "Spacious and modern flat"
                },
                "latitude": {
                    "type": "number",
                    "maximum": 90,
                    "minimum": -90,
                    "example": 6.4281
                },
                "longitude": {
                    "type": "number",
                    "maximum": 180,
                    "minimum": -180,
                    "example": 3.4219
                },
                "price": {
                    "type": "number",
                    "example": 1500000
                },
                "title": {
                    "type": "string",
                    "maxLength": 500,
                    "minLength": 3,
                    "example": "3-Bed Flat in Lekki"
                },
                "type": {
                    "type": "string",
                    "enum": [
                        "rent",
                        "sale",
                        "shortlet"
                    ],
                    "example": "rent"
                }
            }
        },
        "internal_listing_handler.DataListResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/internal_listing_handler.ListingResponse"
                    }
                },
                "pagination": {
                    "$ref": "#/definitions/github_com_davidakpele_property-marketplace_pkg_pagination.Meta"
                }
            }
        },
        "internal_listing_handler.DataResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/internal_listing_handler.ListingResponse"
                }
            }
        },
        "internal_listing_handler.ErrorResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string",
                    "example": "listing not found"
                }
            }
        },
        "internal_listing_handler.ListingResponse": {
            "type": "object",
            "properties": {
                "address": {
                    "type": "string",
                    "example": "14 Admiralty Way, Lekki Phase 1"
                },
                "agent_id": {
                    "type": "string",
                    "example": "a1000000-0000-0000-0000-000000000001"
                },
                "bedrooms": {
                    "type": "integer",
                    "example": 3
                },
                "created_at": {
                    "type": "string"
                },
                "description": {
                    "type": "string",
                    "example": "Spacious and modern flat"
                },
                "id": {
                    "type": "string",
                    "example": "b1000000-0000-0000-0000-000000000001"
                },
                "latitude": {
                    "type": "number",
                    "example": 6.4281
                },
                "longitude": {
                    "type": "number",
                    "example": 3.4219
                },
                "price": {
                    "type": "number",
                    "example": 1500000
                },
                "title": {
                    "type": "string",
                    "example": "3-Bed Flat in Lekki"
                },
                "type": {
                    "type": "string",
                    "example": "rent"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "internal_listing_handler.UpdateListingRequest": {
            "type": "object",
            "required": [
                "address",
                "agent_id",
                "latitude",
                "longitude",
                "price",
                "title",
                "type"
            ],
            "properties": {
                "address": {
                    "type": "string",
                    "maxLength": 500,
                    "minLength": 3,
                    "example": "14 Admiralty Way, Lekki Phase 1"
                },
                "agent_id": {
                    "type": "string",
                    "example": "a1000000-0000-0000-0000-000000000001"
                },
                "bedrooms": {
                    "type": "integer",
                    "minimum": 0,
                    "example": 3
                },
                "description": {
                    "type": "string",
                    "example": "Updated description"
                },
                "latitude": {
                    "type": "number",
                    "maximum": 90,
                    "minimum": -90,
                    "example": 6.4281
                },
                "longitude": {
                    "type": "number",
                    "maximum": 180,
                    "minimum": -180,
                    "example": 3.4219
                },
                "price": {
                    "type": "number",
                    "example": 1800000
                },
                "title": {
                    "type": "string",
                    "maxLength": 500,
                    "minLength": 3,
                    "example": "Updated 3-Bed Flat"
                },
                "type": {
                    "type": "string",
                    "enum": [
                        "rent",
                        "sale",
                        "shortlet"
                    ],
                    "example": "sale"
                }
            }
        },
        "internal_listing_handler.ValidationErrorResponse": {
            "type": "object",
            "properties": {
                "errors": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/internal_listing_handler.ValidationFieldError"
                    }
                }
            }
        },
        "internal_listing_handler.ValidationFieldError": {
            "type": "object",
            "properties": {
                "field": {
                    "type": "string",
                    "example": "price"
                },
                "message": {
                    "type": "string",
                    "example": "must be greater than 0"
                }
            }
        }
    },
...
}